+++ infra/whats-new/api/src/commonMain/kotlin/app/campfire/whatsnew/api/WhatsNewRepository.kt
@@ -1,0 +1,11 @@
+package app.campfire.whatsnew.api
+
+import kotlinx.coroutines.flow.Flow
+
+interface WhatsNewRepository {
+
+  suspend fun getChangelog(): Changelog
+
+  fun observeShouldShowWhatsNew(): Flow<Boolean>
+  suspend fun dismissWhatsNew()
+}