+++ infra/whats-new/api/src/commonMain/kotlin/app/campfire/whatsnew/api/WhatsNewWidgetProvider.kt
@@ -1,0 +1,13 @@
+package app.campfire.whatsnew.api
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+interface WhatsNewWidgetProvider {
+
+  @Composable
+  fun Content(
+    onClick: () -> Unit,
+    modifier: Modifier,
+  )
+}