+++ infra/whats-new/api/src/commonMain/kotlin/app/campfire/whatsnew/api/screen/ChangelogScreen.kt
@@ -1,0 +1,7 @@
+package app.campfire.whatsnew.api.screen
+
+import app.campfire.common.screens.BaseScreen
+import app.campfire.core.parcelize.Parcelize
+
+@Parcelize
+data object ChangelogScreen : BaseScreen(name = "Changelog")